Sequential Conditional Flow

A sequential conditional flow is a conditional flow inheriting from nn_conditional_flow() that applies a sequence of conditional flows, passing the conditioning input to each flow. It is the analog of torch::nn_sequential() for conditional flows.

Usage

nn_sequential_conditional_flow(...)

Arguments

...

A sequence of conditional flows, or a list of conditional flows.

See also

Examples

flow_model <- nn_sequential_conditional_flow(
  nn_affine_coupling_block(10, 5),
  nn_permutation_flow(10),
  nn_affine_coupling_block(10, 5)
)